working on getSpikes
need to have different getSpikesByStim for each curve type (mostly). how to do?

within curveinfo, need a method that takes a spike table or just the timetamps in seconds?... and converts to spiketimes organized by stimulus (stimindex). subclasses will then overload the method in order to cope with differences in data structure