adding eventsToWrite as option
adding option to change which events get written to .nex file. This is to (1) enable testing/probing of errors in writing the imported .nex file to .plx format within Plexon OFS (offline sorter) and (2) to allow user to control the event entries that are written to the .nex file

added nested function to check inputs.

NEED TO TEST!